refactor(web): extract session-detail-aux from SessionDetail
Move DeferredInteractiveReceipt, SessionDetailAuxControls, and
SessionDetailAuxOverlay into session-detail/session-detail-aux.tsx.
SessionDetail is now a 299-line orchestration shell (down from 1840).
Final sub-domain of CS-13.
